3.8 Terminado

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -42,7 +42,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setDouble(idx, comercial.getComision());
+			ps.setBigDecimal(idx, comercial.getComision());
 			return ps;
 		}, keyHolder);
 
@@ -63,7 +63,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comision"))
+                							  rs.getBigDecimal("comision"))
                 						 	
         );
 		
@@ -84,7 +84,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 								rs.getString("nombre"),
 								rs.getString("apellido1"),
 								rs.getString("apellido2"),
-								rs.getFloat("comision"))
+								rs.getBigDecimal("comision"))
 						, id
 				);
 
@@ -166,7 +166,6 @@ public class ComercialDAOImpl implements ComercialDAO {
 						GROUP BY cliente.id, cliente.nombre
 						ORDER BY cuantia DESC;
 					""";
-
 
 		return  jdbcTemplate.query(query, new BeanPropertyRowMapper<>(ClienteDTO.class), id);
